upgrading accounts with Stripe payement

+++ app/controllers/wikis_controller.rb
@@ -3,16 +3,17 @@ class WikisController < ApplicationController
 
   def index
     @wikis = Wiki.all
+    # authorize @wiki
   end
 
   def show
-
     @wiki = Wiki.find(params[:id])
     authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+
   end
 
   def create
@@ -20,8 +21,9 @@ class WikisController < ApplicationController
     @wiki = Wiki.new
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
+    @wiki.private = params[:wiki][:private]
     @wiki.user = current_user
-
+    authorize @wiki
     if @wiki.save
       flash[:notice] = "Wiki was saved."
       redirect_to [@user, @wiki]
@@ -34,14 +36,16 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
 
   def update
     @wiki = Wiki.find(params[:id])
-    #@wiki.title = params[:wiki][:title]
-    #@wiki.body = params[:wiki][:body]
-    @wiki.assign_attributes(wiki_params)
+    @wiki.title = params[:wiki][:title]
+    @wiki.body = params[:wiki][:body]
+    @wiki.private = params[:wiki][:private]
+    # @wiki.assign_attributes(wiki_params)
 
     if @wiki.save
       flash[:notice] = "Wiki was updated."
@@ -56,8 +60,9 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
     if @wiki.destroy
-      flash[:notice] = "The wiki:\"#{@wiki.title}\" has been deleted"
+      flash[:notice] = "The wiki:\"#{@wiki.id}\" has been deleted"
       redirect_to action: 'index'
     else
       flash[:alert] = "There was an error deleting the wiki"
@@ -67,6 +72,6 @@ class WikisController < ApplicationController
 
   private
   def wiki_params
-    params.require(:wiki).permit(:title, :body)
+    params.require(:wiki).permit(:title, :body, :private)
   end
 end
